+++ config/server.ts
@@ -4,4 +4,8 @@ export default ({ env }) => ({
   app: {
     keys: env.array('APP_KEYS'),
   },
+  cron: {
+    enabled: true,
+    tasks: require('./cron-tasks').default({ env }),
+  },
 });